contacts_api_project/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
django-filter>=2.4,<3.0
python-dotenv>=0.19,<1.0
djangorestframework-simplejwt>=4.6,<5.0